requirements.txt
anyio==4.0.0
certifi==2023.7.22
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
notion-client==2.1.0
python-dotenv==1.0.0
sniffio==1.3.0
